+++ packages/manager/modules/pci-universe-components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.4](https://github.com/ovh/manager/compare/@ovh-ux/ng-ovh-pci-universe-components@1.2.3...@ovh-ux/ng-ovh-pci-universe-components@1.2.4) (2025-12-08)
+
+**Note:** Version bump only for package @ovh-ux/ng-ovh-pci-universe-components
+
+
+
+
+
 ## [1.2.3](https://github.com/ovh/manager/compare/@ovh-ux/ng-ovh-pci-universe-components@1.2.2...@ovh-ux/ng-ovh-pci-universe-components@1.2.3) (2025-11-03)
 
 **Note:** Version bump only for package @ovh-ux/ng-ovh-pci-universe-components